# docs/04-接口与数据设计/openapi/internal-api.json
{
  "components": {
    "schemas": {
...
      "ApiResponseObject": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DictionaryItemRequest": {
        "properties": {
          "itemCode": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "DictionaryRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "dictCode": {
            "type": "string"
          },
          "dictName": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "内部用户认证、RBAC、动态菜单、审批任务、系统治理及阶段1主数据接口。",
    "title": "SRM 内部管理端 API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/master-data/organizations/{id}/disable": {
      "post": {
        "operationId": "disable_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用组织",
        "tags": [
          "organization-controller"
        ]
      }
    },
...
    "/api/v1/system/departments/{id}/disable": {
      "post": {
        "operationId": "disable_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用部门",
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/v1/system/dictionaries": {
      "get": {
        "operationId": "list_6",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List dictionary",
        "tags": [
          "dictionary-controller"
        ]
      },
      "post": {
        "operationId": "create_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{dictId}/items": {
      "get": {
        "operationId": "items",
        "parameters": [
          {
            "in": "path",
            "name": "dictId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Execute dictionary",
        "tags": [
          "dictionary-controller"
        ]
      },
      "post": {
        "operationId": "createItem",
        "parameters": [
          {
            "in": "path",
            "name": "dictId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create dictionary item",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{dictId}/items/{itemId}": {
      "put": {
        "operationId": "updateItem",
        "parameters": [
          {
            "in": "path",
            "name": "dictId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update dictionary item",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{dictId}/items/{itemId}/disable": {
      "post": {
        "operationId": "disableItem",
        "parameters": [
          {
            "in": "path",
            "name": "dictId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable dictionary item",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{dictId}/items/{itemId}/enable": {
      "post": {
        "operationId": "enableItem",
        "parameters": [
          {
            "in": "path",
            "name": "dictId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Enable dictionary item",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{id}": {
      "get": {
        "operationId": "get_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get dictionary",
        "tags": [
          "dictionary-controller"
        ]
      },
      "put": {
        "operationId": "update_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{id}/disable": {
      "post": {
        "operationId": "disable_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/v1/system/dictionaries/{id}/enable": {
      "post": {
        "operationId": "enable_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Enable dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/api/v1/system/document-templates/{id}/disable": {
      "post": {
        "operationId": "disable_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
...
    "/api/v1/system/number-rules/{id}/disable": {
      "post": {
        "operationId": "disable_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
...
    "/api/v1/system/positions/{id}/disable": {
      "post": {
        "operationId": "disable_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用岗位",
        "tags": [
          "position-controller"
        ]
      }
    },
...
    "/api/v1/system/roles/{id}/disable": {
      "post": {
        "operationId": "disable_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用角色",
        "tags": [
          "role-controller"
        ]
      }
    },
...
    "/api/v1/system/users/{id}/disable": {
      "post": {
        "operationId": "disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用用户",
        "tags": [
          "user-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "通过当前站点的同源网关访问",
      "url": "/"
    }
  ]
}